SVG Set: Add experimental icon copy tool.
Like we have an example page for Noticons to simplify finding and copy/pasting the right icon, so would be useful for Gridicons.

This folder here is the beginning of what can hopefully become that for an inline SVG set. It uses the code from http://css-tricks.com/svg-use-external-source/ to ensure that SVGs have the benefits of being both inline (so CSS colorable), and referencing an extermal cachable SVG file.

We'll need a grunt script to clean up and batch all the individual CSS files together into one super sprite (some work for this exists already), and we need a CDN to store this on. But so far this example page seems to be working. Let's test it on all devices before we go further.