Attempt to redesign the "Big 4" icons
- New "My Sites" icon: the WordPress logo with a slightly thicker stem, 1.5dp
- New "Reader" icon: facing, flat, no faux 3D, makes it feel more consistent with the others
- New "Me" icon: designed by @melchoyce in #11, presented here with a floating head instead
- Slightly tweaked "Notifications" icon, adjusted for size and position, should be slightly crisper

With regards to the surrounding circle on the WordPress logo (and also the Me for consistency), the working recommendation is 2dp stroke width. Particularly with regards to the logo we're technically in fauxgo country, so that just felt wrong. Even like this it feels a little imbalanced since it makes the w inside look too thin compared to its surroundings.