Grunt: Update new process to put invisible rectangle first
Previously the invisible rectangle was added to the end of the SVG file, which ironically puts the layer topmost in the rendered SVG. Since it's invisible it doesn't actually affect the visual SVG anywhere. But it did affect a common Sketch workflow.

When you dragged the old generated SVG files into Sketch, command clicked to get to the icon itself, you'd get the topmost layer (the invisible square), and couldn't easily change the color of that.

With these new generated SVGs, you can command click and focus will be set on the iconfile.