Add stub PHP plugin of Gridicons for use with WordPress
This is the very first version, very likely needs some work.

This PR basically copies the technique of the React component, and outputs a folder called `php`. Inside is `gridicons.php`, which can be placed in `mu-plugins`. After that, you can get a gridicion output using `echo get_gridicion( 'gridicons-mail';`.)