Build process: Add `<g>` group to SVGs.
It seems an inconspicuous little tag addition. But it is extremely useful as we are able to target it in CSS. With the `g` tag included, we can target both SVG and g tags  separately. This is helpful specifically for the "offset" fix that uses translate to align 18px sized gridicons on subpixels, and allows us to do this in a way that doesn't affect translate tags on the svg element itself.