New icon: Add "popout" icon
This icon is supposed to represent the idea that when clicked, the content you're editing is opened up in a separate window. For example, you might be editing a post, and wanting to edit another post at the same time. Click the popout icon to open a new tab (if in browser) or window (if in app).