Add PDF creation script
This PR includes:

- A new shellscript that utilizes CairoSVG to convert SVG to PDF losslessly
- Readme updates to show instructions for how to run this script
- A folder full of processed PDFs

We'd like PDF versions of every gridicon in order to utilize these in XCode, which lets us easily recolor them after the fact. I've looked at many different SVG to PDF converters, including Inkscape and Batik, but those all came with a bunch of downsides. CairoSVG seems perfect, as it creates highly optimized low-filesize PDF icons.

The shellscript is fairly barebones but it works just fine. I wasn't able to integrate it into grunt, as we I can't see a way to add CairoSVG as an npm dependency to the grunt script. But if we just need PDFs, we can just run the script every once in a while and commit any additions.