Allow individual icons to be imported by gridicons/icon
To make this possible:

	import GridiconExternal from 'gridicons/external';

we need to publish the individual icons in the root of our npm package.

This also changes how do we tell npm which files to include from
whitelisting (using the files prop in package.json)
to blacklisting (using the .npmignore file). The reason for that
is that I've tried to whitelist every individual icon by setting
the files prop to "*.js" and then blacklist only the Gruntfile.js
within .npmignore. It seems that in case of conflict the files prop
takes precedence over .npmignore, so the Gruntfile.js was still included.