Rename build/ to react-icons/
Note that this is necessary to prevent the build/ folder
to be included in the npm package.

For some reason, although the build/ folder was blacklisted
in the .npmignore, it was not excluded from the package.
https://docs.npmjs.com/misc/developers#keeping-files-out-of-your-package
The docs don't mention the build/ folder as one of the always
included directories, though.