Use cjs and esm folders to publish the icons
The idea is that we want to prepare Gridicons to be able
to distribute both CommonJS and ECMAScript Modules.

Initially, the idea was to use the directory root to publish
the CJS modules so others could:

	import external from 'gridicons/external';

The main disadvantage of that approach is that, in the future,
when ECMAScript modules become more common that namespaced
will be taken by the CommonJS modules and we won't be able
to change that if we want to be backwards-compatible.

So, instead, the approach we're taking is being agnostic about
what module system the library user wants to use by default. So,
either:

	import external from 'gridicons/cjs/external';
	import external from 'gridicons/esm/external';

will work.