Allow Gridicons library to be imported by component (#283)
The goal of this PR is to prepare Gridicons to be imported by component.

To use the external Gridicon at the moment we do:

    import Gridicon from 'gridicons';
    <Gridicon icon="external" />

After this PR lands, with the new API we could also import the individual component directly:

    import GridiconExternal from 'gridicons/dist/external';
    <GridiconExternal />

By doing so, we're able to ship only the SVGs that the user needs and the size impact of using the Gridicons library in any project would become as minimal as it can be. For example, at the moment, to use the external Gridicon (278 bytes) in any project you'll pull 81K because the Gridicon component contains the whole icon set. By importing the individual component, you'll only pull 1,2K.